entidad SalesOrderLine añadido para tabla intermedia entre productos y orden(+ func actualizadas)

+++ e-commerce-API/Controllers/ProductController.cs
@@ -3,6 +3,7 @@ using e_commerce_API.Data.Entities;
 using e_commerce_API.Models;
 using e_commerce_API.Services.Implementations;
 using e_commerce_API.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace e_commerce_API.Controllers
@@ -35,55 +36,72 @@ namespace e_commerce_API.Controllers
             }
             return Ok(product);
         }
+
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> CreateProduct(ProductDto productForCreation)
         {
-            Product? product = _mapper.Map<Product>(productForCreation);
-            if (product == null)
+            string role = User.Claims.SingleOrDefault(o => o.Type.Contains("role")).Value;
+            if (role == "Admin")
             {
-                return BadRequest();
+                Product? product = _mapper.Map<Product>(productForCreation);
+                if (product == null)
+                {
+                    return BadRequest();
+                }
+                _productService.AddProduct(product);
+
+                await _productService.SaveChangesAsync();
+
+                return CreatedAtRoute(nameof(GetProductById), new { id = product.Id }, product);
             }
-            _productService.AddProduct(product);
-
-            await _productService.SaveChangesAsync();
-
-            return CreatedAtRoute(nameof(GetProductById), new { id = product.Id }, product);
+            return Forbid("Acceso no autorizado");
         }
-        [HttpPut("{id}")]
 
+        [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> UpdateProduct(int id, ProductDto product)
         {
-            Product productToUpdate = _productService.GetProductById(id);
+            string role = User.Claims.SingleOrDefault(o => o.Type.Contains("role")).Value;
+            if (role == "Admin")
+            {
+                Product productToUpdate = _productService.GetProductById(id);
+                if (productToUpdate == null)
+                    return NotFound("Producto no encontrado");
+                if (product == null)
+                    return BadRequest();
 
-            if ( productToUpdate== null)
-                return NotFound("Producto no encontrado");
-            if (product == null)
-                return BadRequest();
+                _mapper.Map(product, productToUpdate);
 
-            _mapper.Map(product, productToUpdate);
+                _productService.UpdateProduct(productToUpdate);
 
-            _productService.UpdateProduct(productToUpdate);
+                await _productService.SaveChangesAsync();
 
-            await _productService.SaveChangesAsync();
-
-            return Ok(productToUpdate);
+                return Ok(productToUpdate);
+            }
+            return Forbid("Acceso no autorizado");
         }
 
         [HttpDelete("{id}")]
-
+        [Authorize]
         public async Task<IActionResult> DeleteProduct(int id)
         {
-            Product productToDelete = _productService.GetProductById(id);
-
-            if (productToDelete == null)
+            string role = User.Claims.SingleOrDefault(o => o.Type.Contains("role")).Value;
+            if (role == "Admin")
             {
-                return NotFound("Producto no encontrado");
-            }
-            _productService.DeleteProduct(productToDelete);
+                Product productToDelete = _productService.GetProductById(id);
 
-            await _productService.SaveChangesAsync();
-            
-            return NoContent();
+                if (productToDelete == null)
+                {
+                    return NotFound("Producto no encontrado");
+                }
+                _productService.DeleteProduct(productToDelete);
+
+                await _productService.SaveChangesAsync();
+
+                return NoContent();
+            }
+            return Forbid("Acceso no autorizado");
         }
 
     }
